syz-cluster: move email sender code to a separate package
This will facilitate code reuse.
