+++ syz-cluster/email-reporter/handler.go
@@ -10,18 +10,18 @@ import (
 	"time"
 
 	"github.com/google/syzkaller/pkg/email"
+
 	"github.com/google/syzkaller/syz-cluster/pkg/api"
 	"github.com/google/syzkaller/syz-cluster/pkg/app"
+	"github.com/google/syzkaller/syz-cluster/pkg/emailclient"
 	"github.com/google/syzkaller/syz-cluster/pkg/report"
 )
-
-type SendEmailCb func(context.Context, *EmailToSend) (string, error)
 
 type Handler struct {
 	reporter    string
 	apiClient   *api.ReporterClient
 	emailConfig *app.EmailConfig
-	sender      SendEmailCb
+	sender      emailclient.Sender
 }
 
 func (h *Handler) PollReportsLoop(ctx context.Context, pollPeriod time.Duration) {
@@ -68,7 +68,7 @@ func (h *Handler) report(ctx context.Context, rep *api.SessionReport) error {
 		// This should never be happening..
 		return fmt.Errorf("failed to render the template: %w", err)
 	}
-	toSend := &EmailToSend{
+	toSend := &emailclient.Email{
 		Subject: "Re: " + rep.Series.Title, // TODO: use the original rather than the stripped title.
 		To:      rep.Cc,
 		Body:    body,
@@ -130,7 +130,7 @@ func (h *Handler) IncomingEmail(ctx context.Context, msg *email.Email) error {
 	if reply == "" {
 		return nil
 	}
-	_, err := h.sender(ctx, &EmailToSend{
+	_, err := h.sender(ctx, &emailclient.Email{
 		To:        []string{msg.Author},
 		Cc:        msg.Cc,
 		Subject:   "Re: " + msg.Subject,
