+++ syz-cluster/email-reporter/handler_test.go
@@ -11,9 +11,12 @@ import (
 	"github.com/google/syzkaller/syz-cluster/pkg/api"
 	"github.com/google/syzkaller/syz-cluster/pkg/app"
 	"github.com/google/syzkaller/syz-cluster/pkg/controller"
+	"github.com/google/syzkaller/syz-cluster/pkg/emailclient"
 	"github.com/google/syzkaller/syz-cluster/pkg/reporter"
 	"github.com/stretchr/testify/assert"
 )
+
+var testEmailConfig = emailclient.TestEmailConfig()
 
 func TestModerationReportFlow(t *testing.T) {
 	env, ctx := app.TestEnvironment(t)
@@ -26,7 +29,7 @@ func TestModerationReportFlow(t *testing.T) {
 	receivedEmail := emailServer.email()
 	assert.NotNil(t, receivedEmail, "a moderation email must be sent")
 	receivedEmail.Body = nil // for now don't validate the body
-	assert.Equal(t, &EmailToSend{
+	assert.Equal(t, &emailclient.Email{
 		To:      []string{testEmailConfig.ModerationList},
 		Subject: "Re: " + testSeries.Title,
 		// Note that InReplyTo and Cc are nil.
@@ -50,7 +53,7 @@ func TestModerationReportFlow(t *testing.T) {
 	receivedEmail = emailServer.email()
 	assert.NotNil(t, receivedEmail, "an email must be sent upstream")
 	receivedEmail.Body = nil
-	assert.Equal(t, &EmailToSend{
+	assert.Equal(t, &emailclient.Email{
 		To:        testSeries.Cc,
 		Cc:        []string{testEmailConfig.ArchiveList},
 		Subject:   "Re: " + testSeries.Title,
@@ -102,7 +105,7 @@ func TestInvalidReply(t *testing.T) {
 		assert.NoError(t, err)
 		reply := emailServer.email()
 		assert.NotNil(t, reply)
-		assert.Equal(t, &EmailToSend{
+		assert.Equal(t, &emailclient.Email{
 			To:        []string{"user@email.com"},
 			Cc:        []string{"a@a.com", "b@b.com"},
 			Subject:   "Re: Command",
@@ -137,36 +140,25 @@ func setupHandlerTest(t *testing.T, env *app.AppEnvironment, ctx context.Context
 }
 
 type fakeSender struct {
-	ch chan *EmailToSend
+	ch chan *emailclient.Email
 }
 
 func makeFakeSender() *fakeSender {
 	return &fakeSender{
-		ch: make(chan *EmailToSend, 16),
+		ch: make(chan *emailclient.Email, 16),
 	}
 }
 
-func (f *fakeSender) send(ctx context.Context, e *EmailToSend) (string, error) {
+func (f *fakeSender) send(ctx context.Context, e *emailclient.Email) (string, error) {
 	f.ch <- e
 	return "email-id", nil
 }
 
-func (f *fakeSender) email() *EmailToSend {
+func (f *fakeSender) email() *emailclient.Email {
 	select {
 	case e := <-f.ch:
 		return e
 	default:
 		return nil
 	}
-}
-
-var testEmailConfig = &app.EmailConfig{
-	Name:           "name",
-	DocsLink:       "docs",
-	ModerationList: "moderation@list.com",
-	ArchiveList:    "archive@list.com",
-	Sender:         app.SenderSMTP,
-	SMTP: &app.SMTPConfig{
-		From: "a@b.com",
-	},
 }
